+++ src/library/ps/src/ps.h
@@ -32,6 +32,7 @@ SEXP psll_num_threads(SEXP p);
 SEXP psll_cpu_times(SEXP p);
 SEXP psll_memory_info(SEXP p);
 SEXP psll_memory_uss(SEXP p);
+SEXP psll_memory_maxrss(SEXP p);
 SEXP psll_send_signal(SEXP p, SEXP sig);
 SEXP psll_suspend(SEXP p);
 SEXP psll_resume(SEXP p);